Implement a crude backfill for fill(previous)
The previous value will be seeded with the interval right before the
current interval so that, if the first value is missing, the previous
value gets used instead of a null.

This is still a bit crude because `fill(previous)` really should go back
as far back in time as it needs to to find an appropriate value. The
`derivative(value)` solution didn't need this since the fill iterator
took care of filling in the value if it was missing, but we can't rely
on the fill iterator doing heavy lifting when we are the fill iterator.