Return the statement id with each query result
Without the statement id, it is impossible to determine which queries
are related to each other when chunking is used. Since we are now
advocating for people to use chunking because it is better for the
server, we need to support this type of workflow. We now return the
statement id.