Add "partial" boolean to the result and row when the result has been chunked
The "partial" boolean can be attached to both a series inside of a
result and the result itself. When attached to the series, that means
the series has more data and it will be returned as part of the next
result.

The "partial" boolean attached to the result means that there will be
more data for that statement in a future result that can be merged.