Treat null bytes as whitespace in the point parser
We skip null bytes as whitespace in the point parser, but we forget to
treat null bytes as whitespace in other sections of the parser. This
change teaches the parser to treat null bytes as whitespace.

An additional aspect of this though is that null bytes cannot be escaped
so null bytes can never be included in a measurement, tag key, tag
value, or field key. String fields can still have null bytes in their
value, but you won't be able to parse it anywhere else.